+++ lib/test/integration/zmq_test.go
@@ -1,3 +1,5 @@
+// +build ZMQ4
+
 package integration
 
 import (
@@ -5,21 +7,20 @@ import (
 	"time"
 )
 
-var _ = registerIntegrationTest("nanomsg", func(t *testing.T) {
+var _ = registerIntegrationTest("zeromq", func(t *testing.T) {
 	t.Parallel()
 
 	template := `
 output:
-  nanomsg:
+  zmq4:
     urls:
       - tcp://localhost:$PORT
     bind: false
     socket_type: $VAR1
     poll_timeout: 5s
-    max_in_flight: $MAX_IN_FLIGHT
 
 input:
-  nanomsg:
+  zmq4:
     urls:
       - tcp://*:$PORT
     bind: true
@@ -28,7 +29,6 @@ input:
 `
 	suite := integrationTests(
 		integrationTestOpenClose(),
-		integrationTestSendBatch(10),
 		integrationTestStreamParallel(100),
 	)
 	suite.Run(
@@ -38,17 +38,6 @@ input:
 		testOptVarOne("PUSH"),
 		testOptVarTwo("PULL"),
 	)
-	t.Run("with max in flight", func(t *testing.T) {
-		t.Parallel()
-		suite.Run(
-			t, template,
-			testOptSleepAfterInput(500*time.Millisecond),
-			testOptSleepAfterOutput(500*time.Millisecond),
-			testOptVarOne("PUSH"),
-			testOptVarTwo("PULL"),
-			testOptMaxInFlight(10),
-		)
-	})
 	t.Run("with pub sub", func(t *testing.T) {
 		t.Parallel()
 		suite.Run(